preswald/cli.py
import click
import os
import sys
from preswald import __version__ 


# Entry point for CLI
@click.group()
@click.version_option(__version__, message="Preswald CLI Version: %(version)s")
def cli():
    """Preswald CLI - Simplified Data Workflow Tool."""
    pass


# Command: Initialize Project
@cli.command()
@click.argument("project_name")
def init(project_name):
    """
    Initialize a new Preswald project with the necessary folder structure.
    """
    structure = [
        "ingestion/",
        "transformations/",
        "models/",
        "dashboards/",
        "data/",
        "tests/"
    ]
    try:
        os.makedirs(project_name, exist_ok=True)
        for folder in structure:
            os.makedirs(os.path.join(project_name, folder), exist_ok=True)
        click.echo(f"Initialized new Preswald project: {project_name}")
    except Exception as e:
        click.echo(f"Error initializing project: {e}", err=True)


# Command: Run Local Server
@cli.command()
def run():
    """
    Start a local development server for real-time app preview.
    """
    click.echo("Starting local server...")
    # Placeholder: Replace with actual server startup logic
    os.system("python -m http.server 8000")
    click.echo("Server running at http://localhost:8000")


# Command: Execute Pipeline
@cli.command()
@click.argument("pipeline_name")
def pipeline_run(pipeline_name):
    """
    Run a specific data pipeline.
    """
    try:
        click.echo(f"Running pipeline: {pipeline_name}")
        # Placeholder: Replace with actual pipeline execution logic
        pipeline.run_pipeline(pipeline_name)
        click.echo("Pipeline execution completed successfully.")
    except Exception as e:
        click.echo(f"Error executing pipeline: {e}", err=True)
        sys.exit(1)


# Command: Deploy to Vercel
@cli.command()
@click.option("--prod", is_flag=True, help="Deploy to production environment.")
def deploy(prod):
    """
    Deploy the project to Vercel.
    """
    env = "production" if prod else "preview"
    click.echo(f"Deploying project to {env} environment...")
    try:
        os.system("vercel --prod" if prod else "vercel")
        click.echo("Deployment successful!")
    except Exception as e:
        click.echo(f"Deployment failed: {e}", err=True)


# Command: Generate API Endpoint
@cli.command()
@click.argument("model_name")
def api_generate(model_name):
    """
    Generate a shareable API endpoint for a given data model.
    """
    try:
        click.echo(f"Generating API for model: {model_name}")
        # Placeholder: Replace with actual logic
        api_url = f"http://localhost:8000/api/{model_name}"
        click.echo(f"API endpoint generated: {api_url}")
    except Exception as e:
        click.echo(f"Error generating API: {e}", err=True)


# Command: Debug Pipelines
@cli.command()
def debug():
    """
    Launch interactive debugging mode.
    """
    click.echo("Starting interactive debug mode...")
    # Placeholder: Replace with interactive debugging logic
    click.echo("Debugging pipelines interactively...")


# Main CLI Execution
def main():
    try:
        cli()
    except Exception as e:
        click.echo(f"An unexpected error occurred: {e}", err=True)
        sys.exit(1)


if __name__ == "__main__":
    main()
